step 생성
